refactor: extract main.py into robits/core/ package
Split the 3111-line main.py into focused sub-modules under robits/core/:
- io.py: TeeStream
- tools.py: ToolRegistry, ToolDefinition, role_can_use_tool, validators
- context.py: agent_runtime_context, format helpers, tool result delivery
- lifecycle.py: LifecycleEvent, Alarm, lifecycle state machine, alarms
- providers.py: ChatCompletionsProvider, ResponsesProvider, retry logic
- tool_functions.py: workspace, memory, builtin, and agent tools
- roles.py: Role hierarchy, interact functions, load_tools
- session.py: Session, TranscriptEntry, RuntimeEventStream, scheduler

main.py is now a ~230-line harness that owns all globals and re-exports
everything from the sub-modules so existing import main; main.X usage
continues to work. All 260 tests pass.
