+++ robits/core/__init__.py
@@ -1,0 +1,1 @@
+"""Robits core application package."""
